deploy/ArgoCD and shared/ArgoCD - move the deploy specific docs, defaults, required keys, and validation from shared to deploy
